Backend/requirements.txt

aiohttp==3.11.12
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.1
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
coloredlogs==15.0.1
Cython==3.0.11
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Session==0.8.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-auth==2.34.0
googleapis-common-protos==1.64.0
gotrue==2.11.3
greenlet==3.0.3
grpcio==1.66.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
hyperframe==6.1.0
idna==3.8
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.0.0
importlib_resources==6.4.4
inspect2==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.14
langchain-chroma==0.1.3
langchain-community==0.2.12
langchain-core==0.2.35
langchain-ollama==0.1.1
langchain-text-splitters==0.2.2
langsmith==0.1.104
llm-axe==1.1.9
llvmlite==0.43.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
mongopy==0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.8.4
networkx==3.3
nose==1.3.7
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.3.1
onnxruntime==1.19.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
postgrest==0.19.3
posthog==3.5.2
propcache==0.2.1
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.9.1
pypdf==5.3.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.3.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.6
SQLAlchemy==2.0.32
srt==3.5.3
starlette==0.38.2
storage3==0.11.3
streamlit==1.38.0
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.2.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
trio==0.29.0
trio-websocket==0.11.1
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vosk==0.3.45
watchdog==4.0.2
watchfiles==0.23.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.4
yarl==1.18.3
zipp==3.20.1
